enhancement(Updates): Enable for funds

+++ server/graphql/common/update.js
@@ -1,29 +1,35 @@
 import { get } from 'lodash';
 
-import { mustHaveRole } from '../../lib/auth';
 import { purgeCacheForCollective } from '../../lib/cache';
 import models from '../../models';
-import { NotFound, ValidationFailed } from '../errors';
+import { Forbidden, NotFound, Unauthorized, ValidationFailed } from '../errors';
 import { idDecode, IDENTIFIER_TYPES } from '../v2/identifiers';
 
-function require(args, path) {
-  if (!get(args, path)) {
-    throw new ValidationFailed(`${path} required`);
-  }
+function requireArgs(args, paths) {
+  paths.forEach(path => {
+    if (!get(args, path)) {
+      throw new ValidationFailed(`${path} required`);
+    }
+  });
 }
 
 export async function createUpdate(_, args, req) {
+  if (!req.remoteUser) {
+    throw new Unauthorized('You must be logged in to create an update');
+  }
+
   let CollectiveId = get(args, 'update.collective.id');
   if (!CollectiveId) {
     CollectiveId = get(args, 'update.account.legacyId');
   }
-  mustHaveRole(req.remoteUser, 'ADMIN', CollectiveId, 'create an update');
-  require(args, 'update.title');
 
+  requireArgs(args, ['update.title', 'update.html']);
   const collective = await models.Collective.findByPk(CollectiveId);
 
   if (!collective) {
     throw new Error('This collective does not exist');
+  } else if (!req.remoteUser.isAdminOfCollective(collective)) {
+    throw new Forbidden("You don't have sufficient permissions to create an update");
   }
 
   const update = await models.Update.create({
@@ -55,34 +61,66 @@ async function fetchUpdate(id) {
 }
 
 export async function editUpdate(_, args, req) {
-  require(args, 'update.id');
+  if (!req.remoteUser) {
+    throw new Unauthorized('You must be logged in to edit this update');
+  }
+
+  requireArgs(args, ['update.id']);
   let update = await fetchUpdate(args.update.id);
-  update = await update.edit(req.remoteUser, args.update);
   const collective = await models.Collective.findByPk(update.CollectiveId);
+  if (!req.remoteUser.isAdminOfCollective(collective)) {
+    throw new Forbidden("You don't have sufficient permissions to edit this update");
+  }
+
+  update = await update.edit(req.remoteUser, args.update);
   purgeCacheForCollective(collective.slug);
   return update;
 }
 
 export async function publishUpdate(_, args, req) {
+  if (!req.remoteUser) {
+    throw new Unauthorized('You must be logged in to publish this update');
+  }
+
   let update = await fetchUpdate(args.id);
-  update = await update.publish(req.remoteUser, args.notificationAudience);
   const collective = await models.Collective.findByPk(update.CollectiveId);
+  if (!req.remoteUser.isAdminOfCollective(collective)) {
+    throw new Forbidden("You don't have sufficient permissions to publish this update");
+  }
+
+  update = await update.publish(req.remoteUser, args.notificationAudience);
   purgeCacheForCollective(collective.slug);
   return update;
 }
 
 export async function unpublishUpdate(_, args, req) {
+  if (!req.remoteUser) {
+    throw new Unauthorized('You must be logged in to unpublish this update');
+  }
+
   let update = await fetchUpdate(args.id);
-  update = await update.unpublish(req.remoteUser);
   const collective = await models.Collective.findByPk(update.CollectiveId);
+  if (!req.remoteUser.isAdminOfCollective(collective)) {
+    throw new Forbidden("You don't have sufficient permissions to unpublish this update");
+  }
+
+  update = await update.unpublish(req.remoteUser);
   purgeCacheForCollective(collective.slug);
   return update;
 }
 
 export async function deleteUpdate(_, args, req) {
+  if (!req.remoteUser) {
+    throw new Unauthorized('You must be logged in to delete this update');
+  }
+
   let update = await fetchUpdate(args.id);
-  update = await update.delete(req.remoteUser);
   const collective = await models.Collective.findByPk(update.CollectiveId);
+  if (!req.remoteUser.isAdminOfCollective(collective)) {
+    throw new Forbidden("You don't have sufficient permissions to delete this update");
+  }
+
+  update = await update.delete(req.remoteUser);
   purgeCacheForCollective(collective.slug);
   return update;
 }
